implement model, service, add jackOpstions
Implemented the model for user and userList objects which are resutls of the GithubUserRestService. This service has two interfaces implemented. First to search using a search term and second to get user via user name. It is used by UserReopsitoryImpl to search for users and takes advantage of retrofit2 to handle http as well as returning an rxAdroid Observable. The Observable is used by UserReopsitoryImpl to inevitably get and return an Observable list of User class instances.
Also, updated grade with jackOptions.